Revert "specify name, url and index_prefix in repo config"
This reverts commits that add a separate index_prefix config field.

+++ springboard/tools/commands/import_content.py
@@ -1,11 +1,9 @@
-import os
-
+from itertools import chain
 from ConfigParser import ConfigParser
 
 import yaml
-from slugify import slugify
 
-from springboard.utils import parse_repo_name, config_dict
+from springboard.utils import parse_repo_name
 from springboard.tools.commands.bootstrap import BootstrapTool
 from springboard.tools.commands.base import (
     SpringboardToolCommand, CommandArgument)
@@ -40,26 +38,19 @@ class ImportContentTool(BootstrapTool):
             '-n', '--name',
             dest='repo_name',
             help='Give the repository a custom name on disk.'),
-        CommandArgument(
-            '-p', '--index-prefix',
-            dest='index_prefix',
-            help='Use a custom index prefix for the repository on ES.')
     )
 
     def run(self, config, verbose, clobber, repo_dir, repo_url,
-            ini_config, ini_section, update_config, repo_name,
-            index_prefix):
+            ini_config, ini_section, update_config, repo_name):
         config_file, config_data = config
         repo_name = repo_name or parse_repo_name(repo_url)
-        workdir = os.path.join(repo_dir, repo_name)
-        index_prefix = index_prefix or slugify(repo_name)
-        self.clone_repo(workdir=workdir,
-                        repo_url=repo_url,
-                        clobber=clobber,
-                        verbose=verbose)
+        workdir, _ = self.clone_repo(repo_name=repo_name,
+                                     repo_url=repo_url,
+                                     repo_dir=repo_dir,
+                                     clobber=clobber,
+                                     verbose=verbose)
         self.bootstrap(
             workdir,
-            index_prefix,
             config_data.get('models', {}).items(),
             clobber=clobber,
             verbose=verbose)
@@ -70,10 +61,7 @@ class ImportContentTool(BootstrapTool):
         repositories = config_data.setdefault('repositories', {})
 
         if repo_name not in repositories:
-            repositories[repo_name] = {
-                'url': repo_url,
-                'index_prefix': index_prefix
-            }
+            repositories[repo_name] = repo_url
 
             with open(config_file, 'w') as fp:
                 yaml.safe_dump(config_data,
@@ -81,22 +69,22 @@ class ImportContentTool(BootstrapTool):
             self.emit('Added %s to the %s config file.' % (
                 repo_name, config_file))
 
-        config_key = 'unicore.content_repos'
+        config_key = 'unicore.content_repo_urls'
 
         cp = ConfigParser()
         cp.read(ini_config)
         if not cp.has_section(ini_section):
             cp.add_section(ini_section)
 
-        existing_repos = (cp.get(ini_section, config_key)
-                          if cp.has_option(ini_section, config_key)
-                          else '')
-        existing_repos = config_dict(existing_repos)
-        existing_repos[repo_name] = index_prefix
-        cp.set(
-            ini_section, config_key,
-            '\n'.join('%s = %s' % (k, v) for k, v in existing_repos.items()))
-        with open(ini_config, 'w') as fp:
-            cp.write(fp)
-        self.emit(
-            'Updated unicore.content_repos in %s.' % (ini_config,))
+        existing_repo_urls = (cp.get(ini_section, config_key)
+                              if cp.has_option(ini_section, config_key)
+                              else '')
+        existing_repo_urls = existing_repo_urls.strip().split('\n')
+
+        if repo_url not in existing_repo_urls:
+            cp.set(ini_section, config_key,
+                   '\n'.join(chain(existing_repo_urls, [repo_url])))
+            with open(ini_config, 'w') as fp:
+                cp.write(fp)
+            self.emit(
+                'Updated unicore.content_repo_urls in %s.' % (ini_config,))
